Enhance TypeScript configuration and refactor imports to use absolute paths. Added declaration and output directory settings in tsconfig.json. Updated import paths across multiple files for better module resolution. Introduced unit tests for MemoryCache functionality.
