+++ packages/yeti-blue-sdk/src/client/modules/index.ts
@@ -1,1 +1,3 @@
-export * from "./team";
+export * from "@/client/modules/team";
+export * from "@/client/modules/matches";
+export * from "@/client/modules/rankings";
