+++ packages/yeti-blue-sdk/src/fetcher/fetcher.test.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
-import { Fetcher } from "./fetcher";
-import { Cache } from "../cache/types";
+import { Fetcher } from "@/fetcher";
+import { Cache } from "@/cache";
 
 jest.mock("axios");
 const mockedAxios = axios as jest.Mocked<typeof axios>;
