+++ packages/yeti-blue-sdk/src/fetcher/index.ts
@@ -1,1 +1,1 @@
-export { Fetcher } from "./fetcher";
+export { Fetcher } from "@/fetcher/fetcher";
