+++ packages/yeti-blue-sdk/src/index.ts
@@ -1,3 +1,3 @@
-export * from "./client";
-export * from "./cache";
-export type * from "./schemas/types";
+export * from "@/client";
+export * from "@/cache";
+export type * from "@/schemas/types";
