+++ packages/yeti-blue-sdk/src/schemas/index.ts
@@ -1,4 +1,4 @@
-export * from "./team";
-export * from "./event";
-export * from "./match";
-export * from "./rankings";
+export * from "@/schemas/team";
+export * from "@/schemas/event";
+export * from "@/schemas/match";
+export * from "@/schemas/rankings";
