+++ packages/yeti-blue-sdk/tsconfig.json
@@ -3,6 +3,8 @@
     "target": "ES2024",
     "module": "ESNext",
     "moduleResolution": "bundler",
+    "declaration": true,
+    "outDir": "./dist",
     "strict": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
@@ -14,12 +16,5 @@
     }
   },
   "include": ["src"],
-  "exclude": [
-    "node_modules",
-    "dist",
-    "**/*.test.ts",
-    "**/*.spec.ts",
-    "**/__tests__/*",
-    "**/tests/*"
-  ]
+  "exclude": ["node_modules", "dist"]
 }
